move LockRefreshingTimelockService to lock-api
