+++ lock-api/build.gradle
@@ -22,6 +22,11 @@ dependencies {
     processor group: 'org.immutables', name: 'value'
 
     testCompile group: 'junit', name: 'junit'
+    testCompile group: 'org.assertj', name: 'assertj-core'
     testCompile group: 'org.hamcrest', name: 'hamcrest-core'
     testCompile group: 'org.hamcrest', name: 'hamcrest-library'
+    testCompile(group: "org.jmock", name: "jmock", version: libVersions.jmock) {
+      exclude group: 'org.hamcrest'
+    }
+    testCompile group: 'org.mockito', name: 'mockito-core'
 }
